feat: apply technical improvements from dinamo_rent_tr and fix all lint/type errors
Backend:
- DRY: centralize repository helpers (parse_hora, map_fb_error, opt_str) in core::repository
- DRY: auto, reserva, usuario repositories use centralized helpers instead of local duplicates
- Performance: listar_con_filtros() in renta and reserva repositories for backend date filtering
- Performance: async/await + spawn_blocking for heavy commands (renta obtener/cerrar/editar, dashboard)
- Bug fix: update autos.kilometraje on rental close with km_final
- Atomicity: mantenimiento crear/actualizar use with_transaction for atomic INSERT + UPDATE
- Clippy: fix all 16 pre-existing clippy warnings (-D warnings passes clean)

Frontend:
- Modularize api.ts (1286 lines) into 18 domain modules in src/lib/api/
- Calendario page uses backend date filtering instead of client-side
- Add missing pais field to EmpresaConfig type
- Fix costoLavado in test mocks

Infrastructure:
- .gitattributes: force LF line endings for all text files
- .husky/pre-commit: add shebang
- CI: upgrade actions to v7, tauri-action to v1
- tsconfig: exclude test files from tsc (validated by vitest)

Verification:
- cargo clippy -- -D warnings: 0 errors
- tsc --noEmit: 0 errors
- vitest: 267/267 tests pass (30 files)
- vite build: successful

+++ src-tauri/src/services/simit.rs
@@ -561,7 +561,7 @@ pub fn consultar_placa_con_reintentos(
                 CIRCUIT_BREAKER.record_success();
                 metricas.tiempo_captcha_ms += tiempo_captcha;
                 metricas.tiempo_consulta_ms += tiempo_consulta;
-                metricas.reintentos = intento as u32;
+                metricas.reintentos = intento;
                 log::debug!(
                     "Placa {placa} consultada OK: {} registros (intento {}/{}, {}ms)",
                     registros.len(),
@@ -573,7 +573,7 @@ pub fn consultar_placa_con_reintentos(
             }
             Err(e) => {
                 CIRCUIT_BREAKER.record_failure();
-                metricas.reintentos = intento as u32;
+                metricas.reintentos = intento;
                 ultimo_error = Some(e);
                 
                 if intento < max_retries {
